refactor: move handling of multiple yaml files to Serialization.
